BePython/requirements.txt
anyio==3.6.1
click==8.1.3
colorama==0.4.5
fastapi==0.83.0
fastapi-pagination==0.10.0
FastAPI-SQLAlchemy==0.2.1
greenlet==1.1.3
h11==0.13.0
idna==3.3
joblib==1.1.0
numpy==1.23.3
pandas==1.4.4
pydantic==1.10.2
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
scikit-learn==1.1.2
scipy==1.9.1
six==1.16.0
sklearn==0.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.19.1
threadpoolctl==3.1.0
typing_extensions==4.3.0
uvicorn==0.18.3
